pyproject.toml
[tool.poetry]
name = "backend-loginapp"
version = "0.1.0"
description = ""
authors = ["SageChisanga <chisangakabwe16@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.2"
pyjwt = "^2.9.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
jwt = "^1.3.1"
uvicorn = "^0.32.0"
python-jose = "^3.3.0"
pydantic = {extras = ["email"], version = "^2.9.2"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
